requirements-dev.txt
aniso8601==8.0.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
click==7.1.2
cryptography==3.0
decorator==4.4.2
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
pymongo==3.11.0
python-telegram-bot==12.8
pytz==2020.1
six==1.15.0
tornado==6.0.4
Werkzeug==1.0.1
